Renamed components to be more descriptive
