+++ pages/404.page.tsx
@@ -5,7 +5,7 @@ import { Box, Button, Typography } from '@mui/material';
 import { useTranslation } from 'react-i18next';
 import BaseLayout from 'src/components/Layouts/Basic';
 import useGetAppSettings from 'src/hooks/useGetAppSettings';
-import { BoxWrapper } from './styledComponents/BoxWrapper';
+import { StatusPageWrapper } from './styledComponents/StatusPageWrapper';
 
 const Custom404 = (): ReactElement => {
   const { appName } = useGetAppSettings();
@@ -16,7 +16,7 @@ const Custom404 = (): ReactElement => {
         <title>{`404 - Page Not Found | ${appName}`}</title>
       </Head>
 
-      <BoxWrapper boxShadow={3} data-testid="EmptyReport">
+      <StatusPageWrapper boxShadow={3} data-testid="EmptyReport">
         <Box mb={2}>
           <DisabledByDefaultOutlinedIcon fontSize="large" color="disabled" />
         </Box>
@@ -31,7 +31,7 @@ const Custom404 = (): ReactElement => {
             {t('Back to Home')}
           </Button>
         </Box>
-      </BoxWrapper>
+      </StatusPageWrapper>
     </>
   );
 };
