+++ pages/logout.page.tsx
@@ -8,7 +8,7 @@ import { useTranslation } from 'react-i18next';
 import useGetAppSettings from 'src/hooks/useGetAppSettings';
 import { clearDataDogUser } from 'src/lib/dataDog';
 import { ensureSessionAndAccountList } from './api/utils/pagePropsHelpers';
-import { BoxWrapper } from './styledComponents/BoxWrapper';
+import { StatusPageWrapper } from './styledComponents/StatusPageWrapper';
 
 const LogoutPage = ({}): ReactElement => {
   const { t } = useTranslation();
@@ -27,14 +27,14 @@ const LogoutPage = ({}): ReactElement => {
       <Head>
         <title>{`${t('Logout')} | ${appName}`}</title>
       </Head>
-      <BoxWrapper boxShadow={3} data-testid="EmptyReport">
+      <StatusPageWrapper boxShadow={3} data-testid="EmptyReport">
         <Box mb={2}>
           <ExitToAppRoundedIcon fontSize="large" color="disabled" />
         </Box>
         <Typography variant="h5">
           {t("You're currently being logged out.")}
         </Typography>
-      </BoxWrapper>
+      </StatusPageWrapper>
     </>
   );
 };
