+++ src/components/Coaching/CoachingDetail/Activity/Activity.tsx
@@ -17,7 +17,7 @@ import { styled } from '@mui/material/styles';
 import { DateTime } from 'luxon';
 import { useTranslation } from 'react-i18next';
 import AnimatedCard from 'src/components/AnimatedCard';
-import { Header } from 'src/components/Shared/styledComponents/Header';
+import { StyledHeader } from 'src/components/Shared/styledComponents/StyledHeader';
 import {
   Appeal,
   ContactFilterSetInput,
@@ -216,7 +216,7 @@ export const Activity: React.FC<ActivityProps> = ({
     <AnimatedCard>
       <CardHeader
         title={
-          <Header variant="h6">
+          <StyledHeader variant="h6">
             <span>{t('Current Reality')}</span>
             <PeriodText data-testid="ActivityPeriod">
               {formattedDate}
@@ -237,7 +237,7 @@ export const Activity: React.FC<ActivityProps> = ({
               </StyledButton>
             </ButtonGroup>
             <HelpButton articleVar="HELP_URL_COACHING_ACTIVITY" />
-          </Header>
+          </StyledHeader>
         }
       />
       <SectionsContainer>
