+++ src/components/Coaching/CoachingDetail/OutstandingNeeds/OutstandingNeeds.tsx
@@ -20,7 +20,7 @@ import {
   AlignedTableCell,
   ContentContainer,
   LoadMoreButton,
-} from '../StyledComponents';
+} from '../styledComponents/styledComponents';
 import {
   useLoadAccountListCoachingNeedsQuery,
   useLoadCoachingNeedsQuery,
