+++ src/components/Coaching/CoachingDetail/WeeklyReport/WeeklyReport.tsx
@@ -6,10 +6,10 @@ import { DateTime } from 'luxon';
 import { useTranslation } from 'react-i18next';
 import AnimatedCard from 'src/components/AnimatedCard';
 import { MultilineSkeleton } from 'src/components/Shared/MultilineSkeleton';
-import { Header } from 'src/components/Shared/styledComponents/Header';
+import { StyledHeader } from 'src/components/Shared/styledComponents/StyledHeader';
 import { useLocale } from 'src/hooks/useLocale';
 import { dateFormat, dateFormatWithoutYear } from 'src/lib/intlFormat';
-import { ContentContainer } from '../StyledComponents';
+import { ContentContainer } from '../styledComponents/styledComponents';
 import { useWeeklyReportsQuery } from './WeeklyReport.generated';
 
 const CompletedText = styled('span')({
@@ -91,7 +91,7 @@ export const WeeklyReport: React.FC<WeeklyReportProps> = ({
     <AnimatedCard>
       <CardHeader
         title={
-          <Header variant="h6">
+          <StyledHeader variant="h6">
             <span>{t('Weekly Report')}</span>
             <CompletedText data-testid="CompletedText">
               {completedDate}
@@ -112,7 +112,7 @@ export const WeeklyReport: React.FC<WeeklyReportProps> = ({
                 <ChevronRight />
               </StyledButton>
             </ButtonGroup>
-          </Header>
+          </StyledHeader>
         }
       />
       <ContentContainer>
