+++ src/components/Coaching/CoachingDetail/styledComponents/styledComponents.ts
