+++ src/components/Reports/GoalCalculator/CalculatorSettings/Categories/OneTimeGoalsCategory/OneTimeGoalsCategory.tsx
@@ -3,7 +3,7 @@ import { Form, Formik } from 'formik';
 import { useTranslation } from 'react-i18next';
 import * as yup from 'yup';
 import { useGoalCalculator } from '../../../Shared/GoalCalculatorContext';
-import { StyledTypography } from '../../../SharedComponents/styledComponents/StyledTypography';
+import { StyledSectionTitle } from '../../../SharedComponents/styledComponents/StyledSectionTitle';
 import { OneTimeGoalsCategoryForm } from './OneTimeGoalsCategoryForm/OneTimeGoalsCategoryForm';
 
 interface OneTimeGoalsCategoryProps {}
@@ -46,9 +46,9 @@ export const OneTimeGoalsCategory: React.FC<OneTimeGoalsCategoryProps> = () => {
 
   return (
     <>
-      <StyledTypography>
+      <StyledSectionTitle>
         {t('What are your one-time financial goals?')}
-      </StyledTypography>
+      </StyledSectionTitle>
 
       <Formik
         initialValues={initialValues}
