+++ src/components/Reports/GoalCalculator/SharedComponents/GoalCalculatorGrid/GoalCalculatorGrid.tsx
@@ -23,7 +23,7 @@ import * as yup from 'yup';
 import { useLocale } from 'src/hooks/useLocale';
 import { currencyFormat } from 'src/lib/intlFormat';
 import { useGoalCalculator } from '../../Shared/GoalCalculatorContext';
-import { StyledTypography } from '../styledComponents/StyledTypography';
+import { StyledSectionTitle } from '../styledComponents/StyledSectionTitle';
 import { StyledGrid } from './StyledGrid';
 
 const StyledCard = styled(Card)(({ theme }) => ({
@@ -110,7 +110,7 @@ export const GoalCalculatorGrid: React.FC<GoalCalculatorGridProps> = ({
 
   return (
     <>
-      {promptText && <StyledTypography>{t(promptText)}</StyledTypography>}
+      {promptText && <StyledSectionTitle>{t(promptText)}</StyledSectionTitle>}
 
       <Formik
         initialValues={initialValues}
