+++ src/components/Reports/GoalCalculator/SharedComponents/styledComponents/StyledSectionTitle.ts
@@ -1,6 +1,6 @@
 import { Typography } from '@mui/material';
 import { styled } from '@mui/material/styles';
 
-export const StyledTypography = styled(Typography)(({ theme }) => ({
+export const StyledSectionTitle = styled(Typography)(({ theme }) => ({
   marginBottom: theme.spacing(2),
 }));
