+++ src/components/Reports/SavingsFundTransfer/Table/EmptyTable.tsx
@@ -3,7 +3,7 @@ import { HourglassDisabled } from '@mui/icons-material';
 import { Box, Typography } from '@mui/material';
 import { styled } from '@mui/material/styles';
 import { useTranslation } from 'react-i18next';
-import { BoxWrapper } from 'src/components/Shared/styledComponents/BoxWrapper';
+import { EmptyTableWrapper } from 'src/components/Shared/styledComponents/EmptyTableWrapper';
 
 interface Props {
   title: string;
@@ -18,11 +18,11 @@ export const EmptyTable: React.FC<Props> = ({ title, subtitle }) => {
   const { t } = useTranslation();
 
   return (
-    <BoxWrapper boxShadow={3}>
+    <EmptyTableWrapper boxShadow={3}>
       <StyledHourglassIcon fontSize="large" />
       <Typography variant="h5">{t(title)}</Typography>
       <Typography>{t(subtitle)}</Typography>
       <Box sx={{ padding: 1, display: 'flex', gap: 2 }}></Box>
-    </BoxWrapper>
+    </EmptyTableWrapper>
   );
 };
