+++ src/components/Reports/SavingsFundTransfer/styledComponents/StyledFundDataGrid.ts
@@ -1,7 +1,7 @@
 import { styled } from '@mui/material/styles';
 import { DataGrid } from '@mui/x-data-grid';
 
-export const StyledGrid = styled(DataGrid)(({ theme }) => ({
+export const StyledFundDataGrid = styled(DataGrid)(({ theme }) => ({
   '.MuiDataGrid-row:nth-of-type(2n + 1):not(:hover)': {
     backgroundColor: theme.palette.cruGrayLight.main,
   },
