+++ src/components/Settings/Admin/ImpersonateUser/ImpersonateUserAccordion.tsx
@@ -14,7 +14,7 @@ import { AdminAccordion } from 'src/components/Shared/Forms/Accordions/Accordion
 import { AccordionItem } from 'src/components/Shared/Forms/Accordions/AccordionItem';
 import { StyledFormLabel } from 'src/components/Shared/Forms/FieldHelper';
 import { FieldWrapper } from 'src/components/Shared/Forms/FieldWrapper';
-import { StyledBox } from 'src/components/Shared/styledComponents/StyledBox';
+import { PaddedBox } from 'src/components/Shared/styledComponents/PaddedBox';
 import { SubmitButton } from 'src/components/common/Modal/ActionButtons/ActionButtons';
 import useGetAppSettings from 'src/hooks/useGetAppSettings';
 import { getErrorMessage } from 'src/lib/getErrorFromCatch';
@@ -114,7 +114,7 @@ export const ImpersonateUserAccordion: React.FC<
           errors,
         }): ReactElement => (
           <form onSubmit={handleSubmit}>
-            <StyledBox marginTop={4}>
+            <PaddedBox marginTop={4}>
               <FieldWrapper>
                 <TextField
                   required
@@ -132,8 +132,8 @@ export const ImpersonateUserAccordion: React.FC<
                   <FormHelperText error={true}>{errors.user}</FormHelperText>
                 )}
               </FieldWrapper>
-            </StyledBox>
-            <StyledBox marginTop={2}>
+            </PaddedBox>
+            <PaddedBox marginTop={2}>
               <FieldWrapper>
                 <TextField
                   required
@@ -146,7 +146,7 @@ export const ImpersonateUserAccordion: React.FC<
                   onChange={handleChange}
                 />
               </FieldWrapper>
-            </StyledBox>
+            </PaddedBox>
 
             <DialogActions>
               <SubmitButton
