+++ src/components/Settings/Organization/ManageOrganizationAccess/ManageOrganizationAccessAccordion.tsx
@@ -25,7 +25,7 @@ import {
 import { OrganizationAccordion } from 'src/components/Shared/Forms/Accordions/AccordionEnum';
 import { AccordionItem } from 'src/components/Shared/Forms/Accordions/AccordionItem';
 import { FieldWrapper } from 'src/components/Shared/Forms/FieldWrapper';
-import { StyledBox } from 'src/components/Shared/styledComponents/StyledBox';
+import { PaddedBox } from 'src/components/Shared/styledComponents/PaddedBox';
 import { SubmitButton } from 'src/components/common/Modal/ActionButtons/ActionButtons';
 import useGetAppSettings from 'src/hooks/useGetAppSettings';
 import { useLocale } from 'src/hooks/useLocale';
@@ -346,7 +346,7 @@ export const ManageOrganizationAccessAccordion: React.FC<
           }): ReactElement => (
             <form onSubmit={handleSubmit}>
               <h3>{t('Invite someone to administer this organization:')}</h3>
-              <StyledBox marginTop={4}>
+              <PaddedBox marginTop={4}>
                 <FieldWrapper>
                   <TextField
                     required
@@ -369,7 +369,7 @@ export const ManageOrganizationAccessAccordion: React.FC<
                     </FormHelperText>
                   )}
                 </FieldWrapper>
-              </StyledBox>
+              </PaddedBox>
 
               <DialogActions>
                 <SubmitButton
