+++ src/components/Shared/styledComponents/EmptyTableWrapper.ts
@@ -1,7 +1,7 @@
 import { Box } from '@mui/material';
 import { styled } from '@mui/material/styles';
 
-export const BoxWrapper = styled(Box)(({ theme }) => ({
+export const EmptyTableWrapper = styled(Box)(({ theme }) => ({
   backgroundColor: theme.palette.cruGrayLight.main,
   height: 300,
   minWidth: 700,
