+++ src/components/Shared/styledComponents/PaddedBox.ts
@@ -1,6 +1,6 @@
 import { Box } from '@mui/material';
 import { styled } from '@mui/material/styles';
 
-export const StyledBox = styled(Box)(() => ({
+export const PaddedBox = styled(Box)(() => ({
   padding: '0 10px',
 }));
