+++ src/components/Shared/styledComponents/StyledHeader.ts
@@ -1,7 +1,7 @@
 import { Typography } from '@mui/material';
 import { styled } from '@mui/material/styles';
 
-export const Header = styled(Typography)(({ theme }) => ({
+export const StyledHeader = styled(Typography)(({ theme }) => ({
   display: 'flex',
   gap: theme.spacing(2),
   '@media (max-width: 900px)': {
